+++ packages/web3-core/src/types.ts
@@ -32,7 +32,8 @@ export type SupportedProviders<API extends Web3APISpec> =
 	| Web3BaseProvider<API>
 	| LegacyRequestProvider
 	| LegacySendProvider
-	| LegacySendAsyncProvider;
+	| LegacySendAsyncProvider
+	| string;
 
 export type Web3BaseProviderConstructor = new <API extends Web3APISpec>(
 	url: string,