+++ packages/web3-eth-accounts/src/index.ts
@@ -1,3 +1,3 @@
 export * from './account';
 export * from './types';
-export * from './local_storage_wallet';
+export * from './wallet';