+++ packages/web3-eth-ens/src/ens.ts
@@ -1,17 +1,22 @@
+import { Web3Context, SupportedProviders, Web3ContextObject } from 'web3-core';
 import { Address } from 'web3-utils';
-import { RevertInstructionError } from 'web3-common';
+import { RevertInstructionError, EthExecutionAPI } from 'web3-common';
 import { NonPayableCallOptions, TransactionReceipt, Contract } from 'web3-eth-contract';
 import { RESOLVER } from './abi/resolver';
 import { Registry } from './registry';
 import { registryContractAddress } from './config';
 import { Resolver } from './resolver';
 
-export class ENS {
+export class ENS extends Web3Context<EthExecutionAPI> {
 	public registryAddress: string;
 	private readonly _registry: Registry;
 	private readonly _resolver: Resolver;
 
-	public constructor(registryAddr?: string) {
+	public constructor(
+		registryAddr?: string,
+		provider?: SupportedProviders<EthExecutionAPI> | Web3ContextObject,
+	) {
+		super(provider ?? '');
 		this._registry = new Registry(registryAddr);
 		this.registryAddress = registryAddr ?? registryContractAddress; // TODO change this when eth.net is finished
 		this._resolver = new Resolver(this._registry);